backend/requirements.txt
langchain
langchain-community
langchain-docling
langchain-experimental
langchain-qdrant
clip
pillow
torch
sqlalchemy
psycopg2-binary
langchain-groq
langchain-nomic
fastapi-users[sqlalchemy]
passlib[bcrypt]
asyncpg
fastembed-gpu